post-maker/requirements.txt
jinja2
google-generativeai
